# specs/001-rag-chatbot/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  description: |
    API for the Physical AI & Humanoid Robotics Book chatbot with authentication,
    personalization, and Urdu translation features.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.physical-ai-book.com
    description: Production server
...
paths:
...
  /api/chat/conversations:
    get:
      tags: [chat]
      summary: List user conversations
      description: Get all conversations for the authenticated user (returns empty for anonymous)
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Number of conversations to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [chat]
      summary: Create a new conversation
      description: Start a new chat conversation (authenticated or anonymous)
      security:
        - cookieAuth: []
        - {}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  enum: [en, ur]
                  default: en
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /api/chat/conversations/{conversation_id}/messages:
    get:
      tags: [chat]
      summary: Get messages in a conversation
      parameters:
        - in: path
          name: conversation_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'

    post:
      tags: [chat]
      summary: Send a message to the chatbot
      description: Send a user query and receive chatbot response with book citations
      parameters:
        - in: path
          name: conversation_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 10000
                  example: How do I create a ROS2 publisher?
                selected_context:
                  type: object
                  description: Optional text highlighted by user
                  properties:
                    text:
                      type: string
                      maxLength: 2000
                    source_url:
                      type: string
                      format: uri
                    chapter_id:
                      type: string
      responses:
        '200':
          description: Chatbot response
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_message:
                    $ref: '#/components/schemas/Message'
                  assistant_message:
                    $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========== Content Personalization & Translation ==========
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
        language:
          type: string
          enum: [en, ur]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        citations:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Citation'
        created_at:
          type: string
          format: date-time
    Citation:
      type: object
      properties:
        chapter_id:
          type: string
        chapter_title:
          type: string
        section:
          type: string
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        excerpt:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...